requirements.txt
tweepy>=4.12.0
python-dotenv>=0.19.0
requests>=2.26.0
